GameTime now yield time as uint64_t

+++ Common/GameTime.h
@@ -1,9 +1,11 @@
 #ifndef GAME_TIME_H
 #define GAME_TIME_H
 
+#include <stdint.h>
+
 namespace GameTime {
   void initialize();
-  double get();
+  uint64_t get();
 }
 
 #endif
